cve-check: Copy cve-check.bbclass from Poky-warrior
The cve-check class needs to be modified for cve-update-nvd2-native recipe, but
the class file exists in the Poky. So we override the class by copying the file.
We will modify the class in the next commit.

From Poky-warrior rev: 29b5805c629a3d73e4280adc7d109c2aa2b0bdd8